Relocate handlers tests to tests/test_handlers (#588)
