+++ tests/test_handlers.py
@@ -1,0 +1,246 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+""" Unit tests for zeroconf._handlers """
+
+import logging
+import pytest
+import socket
+import time
+import unittest
+import unittest.mock
+
+import zeroconf as r
+from zeroconf import ServiceInfo, Zeroconf
+from zeroconf import const
+
+from . import _clear_cache
+
+log = logging.getLogger('zeroconf')
+original_logging_level = logging.NOTSET
+
+
+def setup_module():
+    global original_logging_level
+    original_logging_level = log.level
+    log.setLevel(logging.DEBUG)
+
+
+def teardown_module():
+    if original_logging_level != logging.NOTSET:
+        log.setLevel(original_logging_level)
+
+
+class TestRegistrar(unittest.TestCase):
+    def test_ttl(self):
+
+        # instantiate a zeroconf instance
+        zc = Zeroconf(interfaces=['127.0.0.1'])
+
+        # service definition
+        type_ = "_test-srvc-type._tcp.local."
+        name = "xxxyyy"
+        registration_name = "%s.%s" % (name, type_)
+
+        desc = {'path': '/~paulsm/'}
+        info = ServiceInfo(
+            type_,
+            registration_name,
+            80,
+            0,
+            0,
+            desc,
+            "ash-2.local.",
+            addresses=[socket.inet_aton("10.0.1.2")],
+        )
+
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        def get_ttl(record_type):
+            if expected_ttl is not None:
+                return expected_ttl
+            elif record_type in [const._TYPE_A, const._TYPE_SRV]:
+                return const._DNS_HOST_TTL
+            else:
+                return const._DNS_OTHER_TTL
+
+        def _process_outgoing_packet(out):
+            """Sends an outgoing packet."""
+            nonlocal nbr_answers, nbr_additionals, nbr_authorities
+
+            for answer, time_ in out.answers:
+                nbr_answers += 1
+                assert answer.ttl == get_ttl(answer.type)
+            for answer in out.additionals:
+                nbr_additionals += 1
+                assert answer.ttl == get_ttl(answer.type)
+            for answer in out.authorities:
+                nbr_authorities += 1
+                assert answer.ttl == get_ttl(answer.type)
+
+        # register service with default TTL
+        expected_ttl = None
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_query(info))
+        zc.registry.add(info)
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_broadcast(info, None))
+        assert nbr_answers == 12 and nbr_additionals == 0 and nbr_authorities == 3
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        # query
+        query = r.DNSOutgoing(const._FLAGS_QR_QUERY | const._FLAGS_AA)
+        assert query.is_query() is True
+        query.add_question(r.DNSQuestion(info.type, const._TYPE_PTR, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.name, const._TYPE_SRV, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.name, const._TYPE_TXT, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.server, const._TYPE_A, const._CLASS_IN))
+        _process_outgoing_packet(zc.query_handler.response(r.DNSIncoming(query.packets()[0]), False))
+        assert nbr_answers == 4 and nbr_additionals == 4 and nbr_authorities == 0
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        # unregister
+        expected_ttl = 0
+        zc.registry.remove(info)
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_broadcast(info, 0))
+        assert nbr_answers == 12 and nbr_additionals == 0 and nbr_authorities == 0
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        expected_ttl = None
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_query(info))
+        zc.registry.add(info)
+        # register service with custom TTL
+        expected_ttl = const._DNS_HOST_TTL * 2
+        assert expected_ttl != const._DNS_HOST_TTL
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_broadcast(info, expected_ttl))
+        assert nbr_answers == 12 and nbr_additionals == 0 and nbr_authorities == 3
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        # query
+        expected_ttl = None
+        query = r.DNSOutgoing(const._FLAGS_QR_QUERY | const._FLAGS_AA)
+        query.add_question(r.DNSQuestion(info.type, const._TYPE_PTR, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.name, const._TYPE_SRV, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.name, const._TYPE_TXT, const._CLASS_IN))
+        query.add_question(r.DNSQuestion(info.server, const._TYPE_A, const._CLASS_IN))
+        _process_outgoing_packet(zc.query_handler.response(r.DNSIncoming(query.packets()[0]), False))
+        assert nbr_answers == 4 and nbr_additionals == 4 and nbr_authorities == 0
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+
+        # unregister
+        expected_ttl = 0
+        zc.registry.remove(info)
+        for _ in range(3):
+            _process_outgoing_packet(zc.generate_service_broadcast(info, 0))
+        assert nbr_answers == 12 and nbr_additionals == 0 and nbr_authorities == 0
+        nbr_answers = nbr_additionals = nbr_authorities = 0
+        zc.close()
+
+    def test_name_conflicts(self):
+        # instantiate a zeroconf instance
+        zc = Zeroconf(interfaces=['127.0.0.1'])
+        type_ = "_homeassistant._tcp.local."
+        name = "Home"
+        registration_name = "%s.%s" % (name, type_)
+
+        info = ServiceInfo(
+            type_,
+            name=registration_name,
+            server="random123.local.",
+            addresses=[socket.inet_pton(socket.AF_INET, "1.2.3.4")],
+            port=80,
+            properties={"version": "1.0"},
+        )
+        zc.register_service(info)
+
+        conflicting_info = ServiceInfo(
+            type_,
+            name=registration_name,
+            server="random456.local.",
+            addresses=[socket.inet_pton(socket.AF_INET, "4.5.6.7")],
+            port=80,
+            properties={"version": "1.0"},
+        )
+        with pytest.raises(r.NonUniqueNameException):
+            zc.register_service(conflicting_info)
+        zc.close()
+
+    def test_register_and_lookup_type_by_uppercase_name(self):
+        # instantiate a zeroconf instance
+        zc = Zeroconf(interfaces=['127.0.0.1'])
+        type_ = "_mylowertype._tcp.local."
+        name = "Home"
+        registration_name = "%s.%s" % (name, type_)
+
+        info = ServiceInfo(
+            type_,
+            name=registration_name,
+            server="random123.local.",
+            addresses=[socket.inet_pton(socket.AF_INET, "1.2.3.4")],
+            port=80,
+            properties={"version": "1.0"},
+        )
+        zc.register_service(info)
+        _clear_cache(zc)
+        info = ServiceInfo(type_, registration_name)
+        info.load_from_cache(zc)
+        assert info.addresses == []
+
+        out = r.DNSOutgoing(const._FLAGS_QR_QUERY)
+        out.add_question(r.DNSQuestion(type_.upper(), const._TYPE_PTR, const._CLASS_IN))
+        zc.send(out)
+        time.sleep(0.5)
+        info = ServiceInfo(type_, registration_name)
+        info.load_from_cache(zc)
+        assert info.addresses == [socket.inet_pton(socket.AF_INET, "1.2.3.4")]
+        assert info.properties == {b"version": b"1.0"}
+        zc.close()
+
+
+def test_ptr_optimization():
+
+    # instantiate a zeroconf instance
+    zc = Zeroconf(interfaces=['127.0.0.1'])
+
+    # service definition
+    type_ = "_test-srvc-type._tcp.local."
+    name = "xxxyyy"
+    registration_name = "%s.%s" % (name, type_)
+
+    desc = {'path': '/~paulsm/'}
+    info = ServiceInfo(
+        type_, registration_name, 80, 0, 0, desc, "ash-2.local.", addresses=[socket.inet_aton("10.0.1.2")]
+    )
+
+    nbr_answers = nbr_additionals = nbr_authorities = 0
+    has_srv = has_txt = has_a = False
+
+    # register
+    zc.register_service(info)
+    nbr_answers = nbr_additionals = nbr_authorities = 0
+
+    # query
+    query = r.DNSOutgoing(const._FLAGS_QR_QUERY | const._FLAGS_AA)
+    query.add_question(r.DNSQuestion(info.type, const._TYPE_PTR, const._CLASS_IN))
+    out = zc.query_handler.response(r.DNSIncoming(query.packets()[0]), False)
+    assert out is not None
+    nbr_answers += len(out.answers)
+    nbr_authorities += len(out.authorities)
+    for answer in out.additionals:
+        nbr_additionals += 1
+        if answer.type == const._TYPE_SRV:
+            has_srv = True
+        elif answer.type == const._TYPE_TXT:
+            has_txt = True
+        elif answer.type == const._TYPE_A:
+            has_a = True
+    assert nbr_answers == 1 and nbr_additionals == 3 and nbr_authorities == 0
+    assert has_srv and has_txt and has_a
+
+    # unregister
+    zc.unregister_service(info)
+    zc.close()
